Added Httparty, and created service for handling the api integration, can now seed db with account balance

+++ db/migrate/20241018111611_create_accounts.rb
@@ -1,7 +1,7 @@
 class CreateAccounts < ActiveRecord::Migration[7.2]
   def change
     create_table :accounts do |t|
-      t.integer :balance
+      t.numeric :balance
 
       t.timestamps
     end
